feat: add dynamics pagination and filter
The pagination is broken in some cases and also needs to reload the view for accesss anothoer pages. In the inital idea, the filter just works with navegator request. Implement this requests to work with ajax could be the best way to achieve a dynamic web site.

This commit add:
 - JSON Response for the class JsonListView, making the pagination work dynamic with the AJAX requests;
 - get_queryset and get_context functions, for easily work with filters and searchs;
 - Dynamic filtering using AJAX requests.